Move initial screen to its own file
